+++ docs/reference/changelog.rst
@@ -21,6 +21,7 @@ Current version
 v5.2.0 - 2022-xx-xx
 -------------------
 
+- ``Added`` Meterstand tonen bij energiecontracten [`#1559 <https://github.com/dsmrreader/dsmr-reader/issues/1559>`_]
 
 
 v5.1.0 - 2022-03-01